requirements.txt
streamlit
pandas
python-dotenv
sqlalchemy
mysql-connector-python